effects: change the `overlayed::Bool` property to `nooverlayed::Bool`
The current naming of `overlayed` is a bit confusing since 
`overlayed === true`, which is the conservative default value, actually
means "it _may_ be overlayed" while `overlayed === false` means 
"this is not absolutely overlayed".
I think it should be named as `nooverlayed`, and then a query name like
`is_nooverlayed` would be more sensible than the alternative 
`is_overlayed`,
which actually should be something like `may_be_overlayed`.